chore(rich-text-editor): move plugins api to its own directory
