+++ packages/picasso/src/utils/index.ts
@@ -60,7 +60,6 @@ export { formatAmount, DEFAULT_CURRENCY, DEFAULT_LOCALE } from './Formatters'
 
 export type { FormatAmount } from './Formatters'
 export * from './get-react-node-text-content'
-export * from './use-memo-object'
 
 export {
   useDeprecationWarning,
